"8:05pm. https://www.reddit.com/r/math/comments/aq0qpd/opinions_on_proof_assistants/
"I think that too much of the discussion about proof assistants gets bogged down in this idea that formal proofs have to be "human-readable" and/or mimic precisely the structure of a handwritten proof. As if in the future proof assistants are just going to be this fancy plugin you can attach to your favorite tex editor and it'll check your proof mathemagically without you having to worry about anything. There was a similar idea in the world of programming and nothing really came of it.

On the contrary, one of the things I like about proof assistants is that often you are forced to structure arguments in different ways that makes them more amenable to formal proof, even if that's not the way you would write them out on paper (that's to say nothing of proof assistants like Coq that additionally demand that you seriously consider the constructivity of your arguments). In this way it offers a fresh intellectual challenge and not just a chore one is obligated to complete out of paranoia.

To make an analogy, it's like if electronic instruments were invented and one guy said "this is great, now we can automate away the playing of acoustic instruments and eliminate human error" and another guy said "no these electronic instruments are worthless because traditional musicians will never learn them and besides, they don't even sound like acoustic instruments anyway". In reality, both guys are missing the point."

"If mathematicians cannot more-or-less straightforwardly translate their (constructive) arguments into Coq, good luck getting them to adapt it. And the other direction is even worse: there is no good reason why Coq proofs shouldn't be human-readable too! It's a consequence of the ugly tactic-based design. Writing Coq is like writing ASM code that generates Haskell code... you get the restrictions of Haskell and the hackiness of ASM at the same time."

"> Writing Coq is like writing ASM code that generates Haskell code...

Perfect analogy. Gonna "borrow" that line."

8:20pm. https://leanprover.github.io/theorem_proving_in_lean/theorem_proving_in_lean.pdf

"For most of its history, mathematics was essentially computational: geometry dealt with constructions of
geometric objects, algebra was concerned with algorithmic solutions to systems of equations, and analysis
provided means to compute the future behavior of systems evolving over time. From the proof of a theorem
to the effect that “for every x, there is a y such that …”, it was generally straightforward to extract an
algorithm to compute such a y given x.

In the nineteenth century, however, increases in the complexity of mathematical arguments pushed mathematicians to develop new styles of reasoning that suppress algorithmic information and invoke descriptions
of mathematical objects that abstract away the details of how those objects are represented. The goal was
to obtain a powerful “conceptual” understanding without getting bogged down in computational details,
but this had the effect of admitting mathematical theorems that are simply false on a direct computational
reading."

8:25pm. "First, we show that the two functions f1 and f2 are equal using function extensionality, and then we cast 0
of type N by replacing f1 by f2 in the type. Of course, the cast is vacuous, because N does not depend on f1.
But that is enough to do the damage: under the computational rules of the system, we now have a closed
term of N that does not reduce to a numeral. "

It might be good to add Lean to the list of things to study.

8:30pm. https://agda.readthedocs.io/en/v2.6.0/language/cubical.html

I am reading this while I have some cake.

8:45pm. Let me back to the game. These quotes might be worth posting."