"1:35pm. Done with chores. It took a lot longer than I thought it would. Holy shit, it is hot outside. Just how can it just suddenly get so hot within a span of a week?
1:40pm. Let me resume.

```
transport : {X : 𝓤 ̇ } (A : X → 𝓥 ̇ ) {x y : X}
          → x ≡ y → A x → A y
transport A (refl x) = 𝑖𝑑 (A x)
```

Now focus me.

...Despite how much practice I got with abstract kinds of programming in Spiral, I am still weak at the type magic done here.

1:45pm. But I think I get this example. `𝑖𝑑 (A x)`. If I think about how `transport` works, what `transport-ap` proves is that `(A ∘ f) x` and `A (f x)` are equal. That is it.

2pm. Focus me focus.

I somehow managed to get to line 2480. The proof I was studying before was only on line 700-800 in the `.agda` file (the one without text) and yet that file has 4.8k lines of code. The intensity code will go up past this point, and so will the exercises arrive. Maybe that will spark my interest a little.

2:10pm.

```
to-Σ-≡ : {X : 𝓤 ̇ } {A : X → 𝓥 ̇ } {σ τ : Σ A}
       → (Σ \(p : pr₁ σ ≡ pr₁ τ) → transport A p (pr₂ σ) ≡ pr₂ τ)
       → σ ≡ τ
to-Σ-≡ (refl x , refl a) = refl (x , a)

from-Σ-≡ : {X : 𝓤 ̇ } {A : X → 𝓥 ̇ } {σ τ : Σ A}
         → σ ≡ τ
         → Σ \(p : pr₁ σ ≡ pr₁ τ) → transport A p (pr₂ σ) ≡ pr₂ τ
from-Σ-≡ (refl (x , a)) = (refl x , refl a)
```

I can't believe I am getting assaulted by boredom already, but I have to bear with it. Let me walk myself through the proofs above.

2:25pm. Let me take a break here. I just focus on this.

```
transport A p (pr₂ σ) ≡ pr₂ τ
```

I sort of get what what this section is trying to teach me - in fact I ran into issues with the type such as there, but I do not understand the concrete details of the above.

It is not really consistent for me. Shouldn't it be possible to write `pr₂ σ ≡ pr₂ τ`?

I'll investigate this after taking a little breather. Boredom is hitting me."