"``` ⁻¹-left∙ : {X : 𝓤 ̇ } {x y : X} (p : x ≡ y)          → p ⁻¹ ∙ p ≡ refl y ⁻¹-left∙ (refl x) = refl (refl x) ```
1:25pm. Focus me focus.

Let me go at it from the top.

```
 p                       ≡⟨ a y p ⟩
 f x (refl x)⁻¹ ∙ f y p  ≡⟨ ap (λ - → (f x (refl x))⁻¹ ∙ -) (κ y p q) ⟩
 f x (refl x)⁻¹ ∙ f y q  ≡⟨ (a y q)⁻¹ ⟩
 q                       ∎
```

I understand everything that is going on here, though not the reason why. At any rate, the sequence of rewrites here is straightforward.

Let me go for the first mystery.

```
  f : (y : X) → x ≡ y → x ≡ y
  f y = collapser (x ≡ y) (c y)
```

This is just the identity. What is this gaining me?

If course if I replace it with the `id` I get a type error in `κ`.

That is the first mystery. I am going to have to get to the bottom of it.

1:40pm.

```
      κ : (y : X) (p q : x ≡ y) → f y p ≡ f y q
      κ y = collapser-wconstancy (x ≡ y) (c y)
```

This one is more straightforward in comparison. I really do not understand why that `f y` is necessary here. In comparison to the rest though, understanding why `f y p ≡ f y q` is more straightforward.

```
      a : (y : X) (p : x ≡ y) → p ≡ (f x (refl x))⁻¹ ∙ f y p
      a x (refl x) = (⁻¹-left∙ (f x (refl x)))⁻¹
```

Now this one is confusing.

1:45pm. In order to gain an understanding of this, what I am going to have to do is inline everything by hand until I reach the point where things click.

But before that, let me read the paper on this theorem."