Let me finally start. I got some rest and have shaken off my early morning grogginess.
I'll do a writeup of the video later, right now let me just resume studying the book.

It feels like I am coming into my own. I am going to learn this thing. Univalent mathematics will be at my beck and call if not in the next month, certainly by the end of the year.

Things would have so much easier for me if instead of tabular or linear Bayesian algorithms I had an algorithm for multiple layers even if it only 2-3. Just that might have been enough for poker.

As it is I have nothing and Spiral will only get me so far.

Really, I should have started work on Spiral only after I've gotten the algorithm, but from a skill development trajectory the path I have taken is the correct one.

Even if it was likely to not bear fruit in the manner that I desired, there is just no way that I could have picked studying the heavily academic type theory right off the bat.

The stuff I am doing now is the beginning of the endgame. Once I figure out the algorithms and the principles behind them, I will be unstoppable.

Doing it this way is what befits me - just sitting back and waiting for it come to me is too cowardly.

Those in the pre-Singularity era who succeeded are supposed to be legendary figures anyway.

This universe might have an ironic streak, but it is not all the way as such.

I am going to have to do a little bit extra to make it hold up its end of the bargain.

Just one little thing to fill in the gap and start the fire.

Today will be another millimeter move closer to getting to that point.