"9:10am. Let me start here. I am not even close to being done with the releases, but I'll leave it for later. Let me read the paper and then I will resume the book.
I'll leave the post for later.

10:10am. Done with the paper. I guess it is a decent survey.

http://www.cs.ru.nl/~herman/

Let me check out the slides by this guy.

While I read them I am going to take a break here as I haven't had dinner yesterday and I am hungry. Also, I've decided on one thing...

```
https://youtu.be/zw6NcwME7yI?t=89
"...Just one anecdote. I once was in the discussion about the future of mathematics and Wisha Groenvel, a famous mathematician was in the audience and in some point he jumped up and went to the blackboard and wrote a large N of mathematics as a function of T and he says what is the limit of T to infinity of mathematics? And he clearly said 'well, it's computers'."
```

I just have no idea what the name of this famous mathematician is supposed to be so I'll cut this quote out. The above is my best translation, but Google does not turn up anything for this or similar names that I have tried.

Google captions are pretty useless for this sort of thing - the system literally translates Grothendieck as 'gross on dick'. It makes other mistakes, but it is especially bad at names."