"8:10am. Let me chill for a while and then I will start.
Yesterday, maybe it was good that job took up my study time since I at least did something. When the turn came to study I turned out to be pretty useless.

It seems it is impossible to underestimate boredom. I just cannot tackle it straight on through sheer willpower. If it was going through the SF volumes, I would have the exercises spaced out for me and I'd be able to learn, but here things are just so arbitrary. It is going to take me a while to learn univalent math, but it will be worth it in the end.

What I am trying to attain here is just so simple - the ability to formalize mathematics. It is not even frontier what I am aiming for, but undergraduate math. This is something that still has not been done to completeness using proof assistants. In a way, what I am trying to do is give myself a math education, except properly. I want to do everything from the ground up - I need this since I lack math talent, and yet need to be the best.

If it is possible to reason through machine learning, then now is the time to find a way of doing that.

https://youtu.be/UTfQwTuri8Y?t=1040
A Fireside Chat with Turing Award Winner Geoffrey Hinton, Pioneer of Deep Learning (Google I/O'19)

Yesterday I watched this talk. It is not like I do not believe in neural nets anymore, but I just do not believe in the experimental approach the field is taking. This cannot be the limit of what is possible to think.

8:25am. Well, then. Let me get back to chilling.

https://www.newton.ac.uk/seminar/20170711100011001
Logic in univalent type theory

When I've shaken off my drowsiness and it is time to resume, I'll start off with this. From there I'll make my way into the courses."