"3:25pm. 20/75. Let me stop this paper here. I want to watch a lecture.
3:30pm. 7m in. These explanations really makes it clearer what they are trying to do. Fibers are just dependent types.

So it makes sense to define it as `f a = b` in that book. What that book was doing is defining inference rules in such a manner.

3:40pm. The structure of these lectures does roughly match the structure of the book.

I think if I could watch a few courses like this, it would give me enough knowledge to just triangulate everything that I need.

3:50pm. Ah, he calls the fibration the function `P : N -> U`.

Ah, that is the fibration. I was not sure whether the definition had a dual meaning till now.

4:05pm. 31m in. This stuff is really losing me.

4:10pm. Let me stop for a bit. I will change direction and watch some homotopy theory videos. There is one by Voevodsky.

I really should find a course for it.

https://github.com/andrejbauer/homotopy-type-theory-course

If I look at the titles for this, all the concepts from HTT draws directly from HT.

4:45pm. https://arxiv.org/pdf/1703.03007.pdf
Homotopy type theory - the logic of space

I'll save this for later. I admit my understanding of types does not have much to do with spaces in geometric sense, but actually having such an understanding would be quite beneficial when dealing with probability distributions. I'll save this paper for later.

I am having no luck finding lectures online.

https://www.youtube.com/playlist?list=PLs9uUZOv0W-2HSTJJ7ec_2pvzuIo_L5LE

Well, there is this, but if feels incomplete. Maybe I should just get a book from Genesis.

At any rate, forget this.

I'll resume the lecture.

Even if I do not know bit and pieces, I'll find a way to fill in the gap.

4:55pm. 38:20.
"If `A` is a discrete space then it is boring. But if A is a circle then it gets really interesting."

This is what gets me about Cubical Agda and such. When I saw them talking about circles, my first thought was to wonder about whether we are talking about the same circle here.

I couldn't even have imagined defining such structures at the type level.

5:25pm. Done with lunch. Let me put in another 30m at least and then I will call it a day.

5:55pm. Let me stop at 65m. Agh, I haven't been able to finish even two lectures in full, but at this point, my focus is starting to snap.

6pm. Tomorrow I will put a decent dent into this. I'll wrap up today's lecture and two more in addition. I will keep doing this until I gain an understanding of this.

The plan was always to just keep going forward in a broad manner. The greater the uncertainty, the greater the need for a larger net.

A year or two down the road, I should have some results from doing this. But even if it a little as just a few months, I should be able to gain a grasp on this."