"8am. I am up and I feel incredibly groggy at the moment, so let me chill for a while. There is nothing much in terms of releases, so I won't have to slack for long.
Today my goal will be to wrap up the last two lectures and then take a look at the other courses. Two days from now, I'll take a little break to compose the half yearly review for the Simulacrum blog.

8:45am. Let me finally start this thing.

8:55am. 10m in.

"The point is, that we are - because our universe is richer in structure we are not killing it off. This structure is playing a role. And in fact, it is playing many roles. When you discover extra structure, one role it plays is that it gets in your way because things that used to work do not work anymore because the structure is there. So sometimes you need to be careful to make sure that it isn't there. But also it gives you opportunity to study new things that you could not study before. So if we just do ordinary algebra we are not going to ask how is this associative, but now we can ask how is this associative and we should discover some sort of homotopicaly richer groups and in fact people have done so."

Yeah, this is in fact exactly that I want. The reason why I cannot into math is because all the languages that I've used up to now have hidden that structure - because of that it created a schism in my mind between programming and mathematics.

This is the right path. If I survive long enough to get to that point, Spiral will play a role in igniting the Singularity, but until then I should just accept that I've pretty much failed at everything and dedicate myself to mathematics.

The only thing one gains in failure is desire, but as little as it is, the desire is the only thing that is important in this world.

9:10am. 16m in. "This is - I find this super cool. Because this is - we are replacing an universal property with a homotopy theoretic notion. We are packing everything - like homotopy there is a slogan in type theory. Type theory eats itself. So in general when you have a formal system and you say that a formal system, that a theory can eat itself that means that the theory can express itself. For instance Godel showed that the Peano arithmetic is rich enough to express the syntax and provability of Peano arithmetic. One of the standard exercises is to formulate type theory inside type theory.

What we are doing here is homotopy theory eats itself. At the end of the day you don't need anything - there is just homotopy theory because you replace category theory with things like 'contractible' and so on. And so now you know and you feel self righteous and you've achieved a goal. There is just homotopy theory. Well, it is not quite homotopy theory, it is homotopy type theory."

9:45am. 46m in. "Of course it is not amazing that we have written this down. We just transcribed ideas that existed already in many shapes and forms - formal, informal, all of them you can find them in category theory, homotopy theory. But what I think is a bit interesting is that it is not taking a lot of technology to get to this point. I am using bare bones formal system with Pis and Sigmas and identity types. I am not building up years and years of topology and analysis to get to this point. I am just doing it directly.

This is important because it tells me if I can make this formal system into a useful tool then I can have this immediate mode of expression where things are not encoded in a complex way. And that always pays off.

Arguments in abstract linear algebra can often be an order of magnitude clearer and shorter than the corresponding thing that is expressed in terms of say matrices and determinants. You'd go crazy. The hope is that its going to be a bit like that. (In homotopy type theory.)"

9:55am. 48m in. Finally the subject turns to higher inductive types. 53m are left in the video. Let me watch this and then I'll take a break.

10am. 54m in. "And this can be very beneficial. For instance, I don't know of any satisfactory construction of real numbers, Cauchy real numbers - without choice predicative, other than the one you get in homotopy type theory."

If I want to crack backprop, I am going to think more carefully about real numbers themselves. So this is of significant interest to me.

10:25am. 1:15h in. I really do not get why having b=b in the second part makes it a circle. I do not get this at all.

But it is not b=b -> S1, but just b=b in the second constructor. That is pretty strange...

Well, only 25m left. Let me finish the video and I will be able to take a break.

11:35am. Done with breakfast. Let me chill for a while longer and then I will resume. That last lecture really got tedious towards the end and I do not feel like resuming right away."