"12:45pm. Let me resume. Finally it is the time for the last lecture. After that I will check out the other courses. Were there any other courses with videos? I think this is the only one...
1:20pm. This quote at 29m might be worth transcribing.

"In practice, when are two structures - when do mathematicians say that two structures are equal? Like two types, two spaces, when do they say that two spaces are equal?

Well actually, they distinguish - well first of all, you put them against the wall and you say 'What is your formal definition!?' You know, you point a gun in their head. Then they are going to say 'Ah, no no, set theory! Set Theory!' And then they will say, 'Two spaces are equal if they have the same elements...' Let's just take the two sets. 'Two sets are equal if they have the same elements.' And then they will say 'Buuut, but! ...Very often I want to consider isomorphic structures.' We've already learned that trick, so equivalent structures. So, they'll often say, 'I want to really consider isomorphic structures.'

And then we'll say 'Ok, but promise to be careful! So when you mean equal you say equal, when you mean isomorphic you say isomorphic.' And they'll go like 'Yeah, yeah, yeah, I promise. Take your gun away.' And they wander off. Ok!

But then, you go and you look at what they do. They don't hold their promise. Because they will easily say 'Ah yeah, but the integer is a subset of the reals. And these other integers here, that is the same thing. We can just mix them freely.' And so then they might say silly things than do not make sense, right? They might say that 4 is an element of the real numbers. But no, but it isn't. 4 is probably an element of the natural numbers, but it is not - if you want it to be an element of the real numbers then that doesn't work.

So they get - from the point of view of formalism they get sloppy. Because they start confusing isomorphism or equivalence on one side and equality on the other. But as anybody who has ever done mathematics knows, this is really, really, really useful and we are not going to give it up. We are just going to say if somebody forces us to say that equality and isomorphism are different, then we are going to say 'Yeah, yeah, yeah!' but when they go away we are going to say 'Ha, ha!' We are just going to continue doing this thing we are doing here because it is so cool.

Ok, now...this is not very mathematical. What we really want is - we want to fix the situation. It is not an acceptable situation. You really want - if something is so useful that you really want it then you should try to fix it.

But it is not so easy. For instance - suppose we just keep doing set theory and then we say 'Ok, let us pretend that equal things are - that isomporphic sets are equal.'. Uhhhh...that is going to give you very strange kind of set theory because what would you be doing? So you cannot just admit that isomorphic things are equal because they are not."

Goes on to talk about proof relevant mathematics and the equivalence axiom. I do not feel like transcribing it all, but it is pretty interesting. This does give me insight about why HoTT is a pretty big deal.

It really is about doing formal mathematics in an informal manner.

This is definitely something I will need.

2pm. 49m in. 46m more of this and then I am done.

...Let me get back into it.

I want to finish it on time today and leave some time for other things.

I won't strain myself to do much else after this lecture.

2:45pm. Got distracted with /a/ for a bit here. Let me finish the lecture. Yeah, I definitely had enough of it.

2:50pm. 1:10h in. I've really been wondering about what `b = b` could ever be used for and here I am finally getting my answer.

3:10pm. 1:25h in. Only ten minutes left. I can't follow this at all, but hopefully some of it will stick into memory.

3:20pm. 1:34h in. "The first time this was done, it was done in a proof assistant because we did not believe ourselves that we were doing it right."

I am interested in how that would work. Do you not need a language with higher inductive types to actually be able to do this kind of reasoning?

3:25pm. Phew, I am finally done.

These lectures are quite hard to watch. But what I know now is barely just the start of what I need in order to do univalent math.

Let me take a look at whether the other courses have recorded sessions.

http://www.andrew.cmu.edu/user/erijke/hott/hott_intro.pdf

Here is another book for this by Edgbert Rijke.

3:30pm. https://www.cs.cmu.edu/~rwh/courses/hott/

Ah, this has a full set of lectures by Harper. 23 to be exact and the first video is a bit over 1h long so I am guessing the rest will be like this. The recording for them is strange - it seems there are two side by side recordings - I guess to save bandwitch the camera tracking the blackboard only takes a few frames per second at a high quality and then there is a low quality recording by the side. At first I thought something was wrong with the video.

It feels a bit off, but I can get used to it.

3:35pm. I won't watch this now. Let me take a look at the book By Rijke. Is it an Agda book or is is standard?

Standard unfortunately. I am really not interested in pure math, so I am not going to be studying from this. It is just not my thing.

Let me check out the HoTT book. It is finally time for me to give it a try."