"4:25pm. Where did the last 50m go?
At any rate, those two lectures that I wanted today sort of spent my energy. Right now I am reading the intro to the HoTT book and what I really want to do right now is study homotopy theory, but I really do not have the energy to start a new thing.

I'll take a look at the first and the homotopy chapters and then call it a day.

4:50pm. https://news.ycombinator.com/item?id=9867465
"Another addendum to the discussion, from an angle more relevant to computer programmers: if we find a computational meaning for the univalence axiom, then it'll be a crazy nuclear weapon of generic programming and code generation.
The univalence axiom says that two types are equal if we can convert their values back and forth (without loss of information). For example, one Boolean type may be defined as an enum containing True and False. But we could rename the values to Foo and Bar and it would be still a fine implementation, because it's easy to convert between the two representations. We can imagine more far-fetched representations too, with more complicated conversion functions.

But in type theory, if two types are equal, then we can freely substitute one for another in any context! For example, if we have a function "Bool -> Bool", and we know that Bool is equal to Bool2, then we can coerce the function to type "Bool2 -> Bool2".

Imagine a complex program that uses a complex piece of data type. If we can provide a correct conversion function to another type (i. e. prove that the type is equivalent to another one), then the univalence axiom can take the conversion function and our program as input, and spits out a new program that uses the new data type, and also preserves all properties of the program. By our current standards of generic programming, this is just insane!

Now, the issue is that currently we don't actually have a computational interpretation for the univalence axiom, so the above magic doesn't yet work. But there is meaningful progress towards that, and there are already some experimental implementations that can do some magic, for example this one:

https://github.com/mortberg/cubicaltt"

"When you think about it we do have a computational meaning and interpretation for the univalence axiom. Your very comment was such an exposition!
What we perhaps don't have is an automated implementation of your exposition. What you're talking about is typecasting where all the terms of one type map to all the terms of another type. And that once shown this the machine could perform type substitution kind of like how type inference works now. (And maybe the machine wouldn't always have to be shown??) Neat idea. Automatic type inference is pretty powerful. Automatic type substitution could be just as powerful perhaps."

5:40pm. https://mathematicswithoutapologies.wordpress.com/2015/05/13/univalent-foundations-no-comment/

I've been reading comments here and it seems HoTT still have some open questions on how to succinctly represent some higher order structures. This along with finding computational expressions of the univalence axioms seem to be the biggest stumbling block as a foundational system.

Well fine. I've certainly seen enough to convince me that it is strictly more expressive than set theory.

6:05pm. http://www.math.columbia.edu/~harris/website/content/10-links/mechanical-mathematicians-duke-university-plum-lecture.pdf

Let me read these slides and then I am done for the day. Tomorrow rather than lectures, I really feel like giving the HoTT book a try. Maybe it won't be too bad."