"4:15pm. 191/490. I'll fill in the gaps using lectures and examples, but just what is that path constructor doing?
Usually the constructors only construct the type itself. Just what if the point of that path?

...Ok, it is `base = base`. So from that I can tell that it connects two cases of the `S^1` type. But in that case what makes it different from a regular function of that same type?

"First of all, the word “generation” should be taken seriously, in the same sense that a group can be freely generated by some set. In particular, because a higher groupoid comes with operations on paths and higher paths, when such an object is “generated” by certain constructors, the operations create more paths that do not come directly from the constructors themselves. For instance, in the higher inductive type S 1 , the constructor loop is not the only nontrivial path from base to base; we have also “loop  loop” and “loop  loop  loop” and so on, as well as loop−1 , etc., all of which are different. This may seem so obvious as to be not worth mentioning, but it is a departure from the behavior of “ordinary” inductive types, where one can expect to see nothing in the inductive type except what was “put in” directly by the constructors."

4:30pm. "Another (temporary) reason for this preference comes from the limitations of existing computer implementations. Proof assistants like COQ and AGDA have ordinary inductive types built in, but not yet higher inductive types. We can of course introduce them by assuming lots of axioms, but this results in only propositional computation rules. However, there is a trick due to Dan Licata which implements higher inductive types using private data types; this yields judgmental rules for point constructors but not path constructors."

4:35pm. 272/490.

"Before turning to the presentation of our results, we briefly review some basic concepts and
theorems from homotopy theory for the benefit of the reader who is not familiar with them. We
also give an overview of the results proved in this chapter"

"Homotopy theory is a branch of algebraic topology, and uses tools from abstract algebra, such
as group theory, to investigate properties of spaces. One question homotopy theorists investigate
is how to tell whether two spaces are the same, where “the same” means homotopy equivalence
(continuous maps back and forth that compose to the identity up to homotopy—this gives the
opportunity to “correct” maps that don’t exactly compose to the identity)."

Wow, this should have been the second chapter.

Yeah, this is what I need no doubt.

https://www.cs.cmu.edu/~rwh/courses/hott/

Now that I've come this far I see that the courses only really cover the first half of the HoTT book.

So maybe I will gain something from the second half.

At any rate, I should watch these lectures after I am done to really have the ideas sink in.

Then I will let them simmer for a while during which I will deal with the rest of the SF volumes.

5:05pm. The HT chapter was a beast so I mostly skimmed it.

319/490. I really did not do jack today, but all this did give me a goal - figure out higher inductive types. They have literally zero precedent in all my experience and if I could figure them out, I will have made a significant stride in understanding HoTT.

"Of the branches of mathematics, category theory is one which perhaps fits the least comfortably in set theoretic foundations. One problem is that most of category theory is invariant under weaker notions of “sameness” than equality, such as isomorphism in a category or equivalence of categories, in a way which set theory fails to capture. But this is the same sort of problem that the univalence axiom solves for types, by identifying equality with equivalence. Thus, in univalent foundations it makes sense to consider a notion of “category” in which equality of objects is identified with isomorphism in a similar way."

But I might as well skim the rest of the book while I am at it.

5:25pm. 385/490. I am really not paying attention to these chapters. This is just heavy math at this point, but I really am interested in reals.

5:45pm. Let me stop here for the day. I just finished lunch and realized that I do not at all feel like reading the book anymore. I had enough of it. The reals will have to wait until I can use higher inductive types in a proof assistant.

Tomorrow to start things off, I will watch the first lecture by Harper and then start writing the review. Last time my thesis was that smart people should concern themselves with making weapons. This time I will be pretty tame in comparison. I will just signal my resolve to challenge the entire field of ML directly. Invention is not working for me, so I'll practice discovery for a while."