"10:20am. https://youtu.be/caSOTjr1z18?t=188
I wonder what language he is using here.

10:50am. https://youtu.be/caSOTjr1z18?t=1590

This could be pretty huge.

10:55am. https://youtu.be/caSOTjr1z18?t=1767

He even mentions partial evaluation here. Yeah, this is pretty interesting.

11:10am. https://youtu.be/caSOTjr1z18?t=2573

Here the thing he is describing that it does not work for is basically Spiral's type system.

11:15am. https://youtu.be/caSOTjr1z18?t=2827

If he can explain the higher inductive type properly, then watching this talk has been worth it.

12:25pm. Done with breakfast. Well, he sort of explained it, but I still have no clear idea of how this is supposed to work. The UA is clear enough to me, but higher inductive types are something entirely else.

https://www.youtube.com/watch?v=O45LaFsaqMA
What if Current Foundations of Mathematics are Inconsistent? | Vladimir Voevodsky

The next video is this one. Let me watch this and then I'll post the 2014 one on the PL sub after proofreading the transcript. I'll also change the essay link to point to the PL thread."