"3:10pm. Did some proofreading of the update and changed a few things.
3:15pm. Right. This milestone is out of the way.

Now what is left would be lectures by Harper. Let me go through those in the next week or so and that will be it as my intro to HoTT. Then I will resume my study of simple type theory.

Every few months, I change. I wonder if I will be able to keep doing this or will I get bogged down in all that logic? Will there at some point be an update saying that I found the algorithm, created the agent and crushed the online gambling lairs?

There is only one way to find that out.

As I've been awake for nearly 9h at this point, let me go through the first lecture and I think I will call it a day. The good thing about all of this is that all the repetition will increase my understanding which will in turn make absorbing different perspectives on this material easier."