"3:45pm. Let me finally start.
3:55pm. https://scs.hosted.panopto.com/Panopto/Pages/Viewer.aspx?id=0945cc7f-48b7-4803-81af-e7193a3f461d

Just how old are these lectures? I expected this to be recent, but he says that they prepared a textbook (the HoTT book) last spring.

He does answer a few question at the 10m mark that I've been wondering by saying that HoTT is self standing and not directly related to algebraic topology. He also mentions proof assistants and somewhat disparages them saying that going through the exercises in them is not the point of this as it does not cultivate the kind of understanding because it is possible to go through the exercises and not learn much.

On the subject of the HoTT book he does says it is not intended for beginners. He does not think it is possible to start learning TT just from that book.

4pm. Oh, the course is 6 years old. I see, I see...

For some reason I thought it would be more recent, but I was wrong.

Oh, yeah, I did mention the patches paper. Well, I'll leave that for later.

4:20pm. 25m in. The lecture is actually pretty good so far. It covers what Andrej Brauer did, but it is thorough. It talks about what type theory is and highlights how proof relevance separates constructivism from other logics.

4:25pm. Oh lol. 27m in.

"...And the proofs of equations are going to correspond - are turn out to abstract the concept of paths in a space which is used as a fundamental tool in homotopy theory and algebraic topology. The idea of path in a space.

Now as a matter of fact, the thing that makes homotopy type theory interesting is the fact that the homotopy theorists never thought of this idea. This never came up. From the point of view of constructive type theory its the most natural thing in the world. And in fact, as a computer scientist I would wager that all of you as computer scientists with some experience in writing, knowing what beautiful code is would have invented this if I'd told you to do it. If I said to you, 'What we are going to do is we are going to study the concept of path in a space, let's get started.' The first thing you would do is would be to write down an abstract type and then you would give the operations associated with that abstract type and start working from there. That's not what you do in the usual formulation of like homotopy theory. Instead what you do is start dropping down to a assembly language and you start hacking stuff up, okay in assembly language.

And its ugly and complicated. But it doesn't need to be. If you do things in kind of type theoretic style everything works out beautifully."

4:35pm. The stuff that he says after that is pretty interesting too but I do not feel like transcribing it all. Maybe I will do it later.

4:40pm. 35m in. Interesting that he put Coq on the right side and states outright that you just need an algorithm.

5pm. 52m in. "What type theory is is a grand unifying theory of computation."

5:05pm. 56m in. "There is no difference between math and code. That is what I would like to see happen."

5:20pm. Only 10m left. At this point my focus is really starting to snap. No way am I doing another lecture today.

I really wasn't sure whether it will be worth it to watch all of them, but it looks like it will be. This first one convinces me.

5:40pm. I am finally done. Let me call it a day here.

This first lecture was actually pretty interesting despite me studying all that other stuff. It flowed pretty naturally.

I should have started here. This is a better resource than either the books or that other recorded course. Unless it suddenly veers off into complicated things, I think by the end of all these lectures I should have good informal understanding of type theory.

5:45pm. Well, I often make mistakes of this sort. For how long had I been studying this by now? 10 days?

Yeah, 10 days exactly was when I finished the last SF exercise.

5:50pm. I'll be done with the lectures by the end of the month at least. I think these are light enough that I could cover them at a pace of 3, maybe 4 per day.

Now let me chill. I've already read Kengan Omega earlier, but Astrid and Iron Ladies are out so let me get to that.

6pm. Manga and games are the only 'work' I will allow after the end line has been crossed. It is actually important to stick to this. I've suffered a lot back in 2016-2017 from trying to squeeze in extra work past during what should be my break time. Stress management is a very real skill that anybody who maintain this pace needs to master.

Some programmers have the opposite problem in that they can never muster the motivation to keep going, but for people like me who have internalized their purpose, the opposite problem - that of stopping tends to be the problem. Playing the willpower card will get called as a bluff past certain point and I am at that point right now.

It is time to rest."